Create Notificiation Function
